Refactoring: moving some template to separate files + classMap fix.
