+++ client/common/lib/elements/configuration-section-header.ts
@@ -7,7 +7,7 @@ import { customElement, property } from 'lit/decorators.js'
 import { LivechatElement } from './livechat'
 
 @customElement('livechat-configuration-section-header')
-export class ConfigurationRowElement extends LivechatElement {
+export class ConfigurationSectionHeaderElement extends LivechatElement {
   @property({ attribute: false })
   public override title: string = 'title'
